Add progress to show on forecast item

+++ weather/presentation/src/main/java/com/jnasser/weather/presentation/weather_detail/model/ForecastDataUi.kt
@@ -5,5 +5,5 @@ data class ForecastDataUi(
     val icon: String,
     val maxTemperature: String,
     val minTemperature: String,
-    val progress: Double
+    val progress: Float
 )
